Typos

+++ src/core/PopIII.h
@@ -18,7 +18,7 @@ extern "C"
   
   void initialize_time_interp_arrays();
   double CCSN_PopIII_Fraction(int Snapshot, int last_snap); 
-  float interp_mass(float lifetime);
+  double interp_mass(double lifetime);
   double IMFnorm(double MminIMF, double MmaxIMF);
   double get_StellarAge(double StarMass);
   double getIMF(double StarMass);
